+++ packages/crayon/src/components/Segment/Segment.stories.tsx
@@ -1,32 +1,33 @@
 import { Meta, StoryFn } from "@storybook/react"
 
+import { Segments } from "./Segments"
 import { Segment } from "./Segment"
-import { SegmentButton } from "./SegmentButton"
 import { css } from "@emotion/react"
 import { useState } from "react"
 
-const meta: Meta<typeof Segment> = {
+const meta: Meta<typeof Segments> = {
   title: "UI/Segment",
-  component: Segment
+  component: Segments
 }
 
 export default meta
 
-const Template: StoryFn<typeof Segment> = (props) => {
-  const [value, setValue] = useState("epic")
+const Template: StoryFn<typeof Segments> = (props) => {
+  const [value, setValue] = useState("segment1")
 
   return (
-    <Segment
+    <div
       css={css`
-        width: 500px;
+        width: 100%;
+        height: 50px;
       `}
-      {...props}
-      currentValue={value}
-      onValueChange={setValue}
     >
-      <SegmentButton value="magazine">magazine</SegmentButton>
-      <SegmentButton value="epic">epic</SegmentButton>
-    </Segment>
+      <Segments {...props} currentValue={value} onValueChange={setValue}>
+        <Segment value="segment1">segment1</Segment>
+        <Segment value="segment2">segment2</Segment>
+        <Segment value="segment3">segment3</Segment>
+      </Segments>
+    </div>
   )
 }
