+++ packages/crayon/src/components/Segment/index.ts
@@ -1,2 +1,2 @@
+export * from "./Segments"
 export * from "./Segment"
-export * from "./SegmentButton"
