+++ packages/crayon/src/components/index.ts
@@ -4,6 +4,7 @@ export * from "./Button"
 export * from "./Checkbox"
 export * from "./Modal"
 export * from "./Ripple"
+export * from "./Segment"
 export * from "./Showcase"
 export * from "./Slider"
 export * from "./Switch"
